Reorganized the functions into the components and implemented routes for the notepad
