+++ public/index.html
@@ -24,7 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Notepad</title>
     <script src="https://accounts.google.com/gsi/client" async defer></script>
   </head>
   <body>
