+++ src/components/NotepadPage.css
@@ -5,6 +5,12 @@
   height: 100vh;
   width: 100%;
 }
+
+.Info {
+  color: black;
+  padding: 10px;
+  text-align: center;
+}
   
 .Notepad-Title {
   text-align: center;
@@ -24,20 +30,7 @@
   margin: 0 15% 0 15%;
   border-left: 3px solid red;
 }
-  
-.Page textarea {
-  font-size: 16px;
-  font-family: 'Segoe UI';
-  line-height: 24px;
-  padding-left: 5px;
-  border: none;
-  outline: none;
-  width: 100%;
-  resize: none;
-  background: transparent;
-  overflow: hidden;
-}
-  
+
 .Notepad-Title input {
   border: none;
   outline: none;
@@ -48,5 +41,18 @@
   width: 100%;
   font-weight: bold;
   background-color: #f5f5f4;
+}
+  
+.Page textarea {
+  font-size: 18px;
+  font-family: 'Segoe UI';
+  line-height: 24px;
+  padding-left: 5px;
+  border: none;
+  outline: none;
+  width: 100%;
+  resize: none;
+  background: transparent;
+  overflow: hidden;
 }
